review-fix: Move Feedback component into AccountsList
